requirements.txt
streamlit==1.38.0
tensorflow-cpu==2.8.0
Pillow==10.4.0
numpy>=1.20